# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Racks API documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/auth/building/{id}": {
            "get": {
                "tags": [
                    "Building"
                ],
                "summary": "Get building",
                "description": "Get building",
                "operationId": "RetrieveBuilding",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Building ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get building",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/RetrieveBuildingResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such building",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/NoSuchBuildingForBuildingResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Building"
                ],
                "summary": "Delete building",
                "description": "Delete building",
                "operationId": "DeleteBuilding",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Building ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Building deleted - No content"
                    },
                    "404": {
                        "description": "No such building",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/NoSuchBuildingForBuildingResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unable to delete building",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/UnableToDeleteBuildingResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Action not allowed for this department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/PermissionExceptionResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Building"
                ],
                "summary": "Update building",
                "description": "Update building",
                "operationId": "UpdateBuilding",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Building ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/UpdateBuildingRequest"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Building updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/BuildingUpdatedResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such building",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/NoSuchBuildingForBuildingResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Building with this name already exists in this site",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/BuildingNameExceptionResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Action not allowed for this department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/PermissionExceptionResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Unable to update building",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/UnableToCreateBuildingResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/auth/department/{id}": {
            "get": {
                "tags": [
                    "Department"
                ],
                "summary": "Get department",
                "description": "Get department",
                "operationId": "RetrieveDepartment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Department ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/RetrieveDepartmentResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No such department",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "$ref": "#/components/schemas/NoSuchDepartmentForDepartmentResponse"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/auth/device/vendors": {
            "get": {
                "tags": [
                    "Device"
                ],
                "summary": "Get device vendors",
                "description": "Get device vendors list",
                "operationId": "GetDeviceVendors",
                "responses": {
                    "200": {
                        "description": "Get device vendors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "oneOf": [
                                                {
                                                    "properties": {
                                                        "item_type": {
                                                            "type": "string",
                                                            "example": "device_vendor"
                                                        },
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "description": "Device vendor",
                                                                "type": "string"
                                                            },
                                                            "example": [
                                                                "Cisco",
                                                                "Huawei",
                                                                "HP"
                                                            ]
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PermissionExceptionResponse": {
                "title": "Action not allowed for this department",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Action not allowed for this department"
                    }
                },
                "type": "object"
            },
...
            "UpdateBuildingRequest": {
                "title": "Update building request",
                "properties": {
                    "name": {
                        "description": "Building name",
                        "type": "string",
                        "example": "Some building",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
            "BuildingNameExceptionResponse": {
                "title": "Building with this name already exists in this site",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Building with this name already exists in this site"
                    }
                },
                "type": "object"
            },
            "BuildingUpdatedResponse": {
                "title": "Update building",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Building name"
                    },
                    "site_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "updated_by": {
                        "type": "string",
                        "example": "Some user"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-28 16:32:21"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-28 16:32:21"
                    }
                },
                "type": "object"
            },
            "NoSuchBuildingForBuildingResponse": {
                "title": "No such building",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No such building"
                    }
                },
                "type": "object"
            },
...
            "RetrieveBuildingResponse": {
                "title": "Get building",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Building name"
                    },
                    "site_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "updated_by": {
                        "type": "string",
                        "example": "Some user"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-28 16:32:21"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-28 16:32:21"
                    }
                },
                "type": "object"
            },
            "UnableToCreateBuildingResponse": {
                "title": "Unable to create building",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unable to create building"
                    },
                    "error": {
                        "type": "string",
                        "example": "Some internal error"
                    }
                },
                "type": "object"
            },
            "UnableToDeleteBuildingResponse": {
                "title": "Unable to delete building",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unable to delete building"
                    },
                    "error": {
                        "type": "string",
                        "example": "Some internal error"
                    }
                },
                "type": "object"
            },
...
            "NoSuchDepartmentForDepartmentResponse": {
                "title": "No such department",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "No such department"
                    }
                },
                "type": "object"
            },
            "RetrieveDepartmentResponse": {
                "title": "Get department",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Department name"
                    },
                    "region_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2024-01-28 16:32:21"
                    },
                    "updated_at": {
                        "type": "string",
                        "example": "2024-01-28 16:32:21"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}